+++ src/lib/global-babel.js
@@ -5,5 +5,8 @@ module.exports = {
     require.resolve('babel-preset-es2015'),
     require.resolve('babel-preset-stage-0'),
     require.resolve('babel-preset-react')
+  ],
+  plugins: [
+    require.resolve('babel-plugin-transform-runtime')
   ]
 };